Move ellipsis.rb into rules directory
