+++ lib/pragmatic_segmenter/process.rb
@@ -2,7 +2,7 @@
 require 'pragmatic_segmenter/list'
 require 'pragmatic_segmenter/abbreviation_replacer'
 require 'pragmatic_segmenter/number'
-require 'pragmatic_segmenter/ellipsis'
+require 'pragmatic_segmenter/rules/ellipsis'
 require 'pragmatic_segmenter/exclamation_words'
 require 'pragmatic_segmenter/punctuation_replacer'
 require 'pragmatic_segmenter/between_punctuation'
