+++ lib/pragmatic_segmenter/rules/ellipsis.rb
